fix(creator): move image profiles to an always-compiled module
The generation module is gated behind the ai feature, but assets.rs and
r2.rs — which call cap_image_bytes — compile in the Community Edition
build, so CI's no-default-features check failed with E0433. Move
RuntimeImageProfile, runtime_image_profile, cap_stored_dims, and
cap_image_bytes (plus their tests) to a new ungated image_profiles
module; generation.rs keeps infer_behavior and calls into it.
